+++ packages/deployments/subgraph/src/amarok-runtime-v0/mapping/transfers.ts
@@ -57,6 +57,7 @@ export function handleXCalled(event: XCalled): void {
   transfer.canonicalId = event.params.params.canonicalId;
   transfer.canonicalDomain = event.params.params.canonicalDomain;
   transfer.asset = getOrCreateAsset(event.params.local).id;
+  transfer.transactingAsset = event.params.asset;
 
   // Message
   let message = OriginMessage.load(event.params.messageHash.toHex());