+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Next Release
 
+## Amarok 0.2.7-alpha.3
+
+- [sdk] more fixes
+
+## Amarok 0.2.7-alpha.2
+
+- [sdk] optimisation for fee caclulation on optimism
+
 ## Amarok 0.2.7-alpha.1
 
 - [sequencer] extra errors